backend/requirements.txt
fastmcp>=2.2.0
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
anthropic>=0.40.0
pydantic>=2.8.0
beautifulsoup4>=4.12.0
ruff>=0.11.0
pypdf>=5.1.0
python-multipart>=0.0.9
